Better.

+++ .github/scripts/build-posix.sh
@@ -53,7 +53,7 @@ echo "::group::Compiling"
 
 cd /tmp/liquidsoap-full
 
-cp PACKAGES.default PACKAGES
+test -f PACKAGES || cp PACKAGES.default PACKAGES
 
 # Workaround
 touch liquidsoap/configure
